feat: 一部の機能のパッケージ見直しとフォーマッター適用
